+++ qcelemental/tests/test_molecule.py
@@ -111,7 +111,7 @@ def test_water_minima_data():
     mol_dict['name'] = "water dimer"
     mol = Molecule(orient=True, **mol_dict)
 
-    assert len(str(mol)) == 661
+    assert len(mol.pretty_print()) == 661
     assert len(mol.to_string("psi4")) == 479
 
     assert sum(x == y for x, y in zip(mol.symbols, ['O', 'H', 'H', 'O', 'H', 'H'])) == mol.geometry.shape[0]
